Refactor sessionspaces update
